Revert "Declare randomseed as an unsigned long to be safe" This reverts commit 07abb929a6ace3d7be745c3a419a1e07a43ed2c7. Original: c5b99c8ee0863c3538332d441a1366a7784cad64

+++ lalapps/src/inspiral/posterior/LALInferenceMCMC.c
@@ -394,7 +394,7 @@ void initializeMCMC(LALInferenceRunState *runState)
     }
 
   INT4 verbose=0,tmpi=0;
-  unsigned long randomseed=0;
+  unsigned int randomseed=0;
   REAL8 trigSNR = 0.0;
   REAL8 tempMin = 1.0;
   REAL8 tempMax = 50.0;
